Canonical CI: grouped-tests.yml + root test/test_groups.toml
Convert the root test workflow (CI.yml) to the canonical thin caller of
SciML/.github grouped-tests.yml@v1, with the version/group/OS matrix declared
once in test/test_groups.toml.

Category B refactor (Aqua/ExplicitImports previously ran inline in
runtests.jl with no separate QA group):
- runtests.jl is now a GROUP dispatcher (default GROUP=All).
- Functional suite moved to test/core_tests.jl, run under GROUP Core/All
  (includes the allocation tests, formerly gated on GROUP all/nopre).
- Aqua + ExplicitImports moved into an isolated QA env
  (test/qa/Project.toml with Aqua/ExplicitImports/Test + Static via
  [sources] path="../..", julia compat "1.10") and test/qa/qa.jl, run only
  under GROUP==QA. No checks were skipped, silenced, or excluded.
- Root Project.toml: dropped Aqua/ExplicitImports from [extras]/[targets]/
  [compat] (now in the QA env); kept AllocCheck + Test for Core.

test_groups.toml: [Core] versions [lts,1,pre] + os [ubuntu-latest,
windows-latest] (preserves the old 2-OS functional matrix); [QA] versions
[lts,1].
