+++ Project.toml
@@ -11,9 +11,7 @@ SciMLPublic = "431bcebd-1456-4ced-9d72-93c2757fff0b"
 
 [compat]
 AllocCheck = "0.1, 0.2"
-Aqua = "0.8.4"
 CommonWorldInvalidations = "1"
-ExplicitImports = "1.9"
 IfElse = "0.1"
 PrecompileTools = "1.2.0"
 SciMLPublic = "1.0.0"
@@ -22,9 +20,7 @@ julia = "1.10"
 
 [extras]
 AllocCheck = "9b6a8646-10ed-4001-bbdc-1d2f46dfbb1a"
-Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
-ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AllocCheck", "Aqua", "ExplicitImports", "Test"]
+test = ["AllocCheck", "Test"]
