+++ test/qa/Project.toml
@@ -1,0 +1,14 @@
+[deps]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
+Static = "aedffcd0-7271-4cad-89d0-dc628f76c6d3"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[sources]
+Static = { path = "../.." }
+
+[compat]
+Aqua = "0.8.4"
+ExplicitImports = "1.9"
+Test = "1"
+julia = "1.10"
