+++ test/test_groups.toml
@@ -1,0 +1,6 @@
+[Core]
+versions = ["lts", "1", "pre"]
+os = ["ubuntu-latest", "windows-latest"]
+
+[QA]
+versions = ["lts", "1"]
